Update guessing game project
This is my implementation of the guessing game project up to the section
on generating random numbers. The difference is I have broken things out
into a number of helper functions.

I also added a win condition, so if the player guesses in MAX_WINNING_TRIES,
they are considered a winner.